Moved into tech namespace from think namespace.
