+++ java/tech/datatype/ArrayView.java
@@ -1,4 +1,4 @@
-package think.datatype;
+package tech.datatype;
 import java.nio.*;
