+++ java/tech/datatype/IntArrayView.java
@@ -1,10 +1,10 @@
-package think.datatype;
+package tech.datatype;
 
 import java.util.Arrays;
 
-public final class ByteArrayView extends ArrayViewBase
+public final class IntArrayView extends ArrayViewBase
 {
-    public final byte[] data;
+    public final int[] data;
     void checkDataLength() throws Exception
     {
 	if ( data.length < (offset + capacity) )
@@ -12,18 +12,18 @@ public final class ByteArrayView extends ArrayViewBase
 		(String.format("data length %s is less than offset %s + capacity %s.",
 			       data.length, offset, capacity));
     }
-    public ByteArrayView( byte[] d, int o, int cap ) throws Exception
+    public IntArrayView( int[] d, int o, int cap ) throws Exception
     {
 	super(o, cap);
 	data = d;
 	checkDataLength();
     }
-    public ByteArrayView( byte[] d, int o )
+    public IntArrayView( int[] d, int o )
     {
 	super(o, d.length - o);
 	data = d;
     }
-    public ByteArrayView( byte[] d )
+    public IntArrayView( int[] d )
     {
 	super(d.length);
 	data = d;
@@ -32,44 +32,44 @@ public final class ByteArrayView extends ArrayViewBase
     /**
        Member function construction to allow chaining from an existing view while preserving type.
      */
-    public final ByteArrayView construct( int offset, int capacity ) throws Exception
+    public final IntArrayView construct( int offset, int capacity ) throws Exception
     {
-	return new ByteArrayView(data, offset, capacity);
+	return new IntArrayView(data, offset, capacity);
     }
 
-    public final byte get(int idx)
+    public final int get(int idx)
     {
 	return data[index(idx)];
     }
-    public final void set(int idx, byte value)
+    public final void set(int idx, int value)
     {
 	data[index(idx)] = value;
     }
-    public final void pluseq(int idx, byte value)
+    public final void pluseq(int idx, int value)
     {
 	data[index(idx)] += value;
     }
-    public final void minuseq(int idx, byte value)
+    public final void minuseq(int idx, int value)
     {
 	data[index(idx)] -= value;
     }
-    public final void multeq(int idx, byte value)
+    public final void multeq(int idx, int value)
     {
 	data[index(idx)] *= value;
     }
-    public final void diveq(int idx, byte value)
+    public final void diveq(int idx, int value)
     {
 	data[index(idx)] /= value;
     }
-    public final void fill(byte value)
+    public final void fill(int value)
     {	
         Arrays.fill(data, offset, (offset + capacity), value);
     }
-    public final ByteArrayView toView(int new_offset, int len) throws Exception
+    public final IntArrayView toView(int new_offset, int len) throws Exception
     {
-	return new ByteArrayView(data, offset + new_offset, len);
+	return new IntArrayView(data, offset + new_offset, len);
     }
-    public final ByteArrayView toView(int offset) throws Exception
+    public final IntArrayView toView(int offset) throws Exception
     {
 	return toView(offset, length() - offset);
     }
